refactor(email_sender): extract class from notifier module
