+++ tests/dag_generator_test.py
@@ -4,7 +4,7 @@
 import pandas as pd
 import pytest
 from dags.ro_dou.dou_dag_generator import merge_results
-from dags.ro_dou.notification.notifier import Notifier, repack_match
+from dags.ro_dou.notification.email_sender import EmailSender, repack_match
 
 
 def test_repack_match(report_example):
@@ -22,53 +22,53 @@ def test_repack_match(report_example):
 
 
 @pytest.fixture
-def notifier(report_example):
-    notifier = Notifier(None)
-    notifier.search_report = report_example
-    return notifier
+def email_sender(report_example):
+    email_sender = EmailSender(None)
+    email_sender.search_report = report_example
+    return email_sender
 
-def test_convert_report_dict__returns_list(notifier):
-    tuple_list = notifier.convert_report_dict_to_tuple_list()
+def test_convert_report_dict__returns_list(email_sender):
+    tuple_list = email_sender.convert_report_dict_to_tuple_list()
     assert isinstance(tuple_list, list)
 
-def test_convert_report_dict__returns_tuples(notifier):
-    tuple_list = notifier.convert_report_dict_to_tuple_list()
+def test_convert_report_dict__returns_tuples(email_sender):
+    tuple_list = email_sender.convert_report_dict_to_tuple_list()
     for tpl in tuple_list:
         assert isinstance(tpl, tuple)
 
-def test_convert_report_dict__returns_tuples_of_seven(notifier):
-    tuple_list = notifier.convert_report_dict_to_tuple_list()
+def test_convert_report_dict__returns_tuples_of_seven(email_sender):
+    tuple_list = email_sender.convert_report_dict_to_tuple_list()
     for tpl in tuple_list:
         assert len(tpl) == 7
 
-def test_convert_report_to_dataframe__rows_count(notifier):
-    df = notifier.convert_report_to_dataframe()
+def test_convert_report_to_dataframe__rows_count(email_sender):
+    df = email_sender.convert_report_to_dataframe()
     # num_rows
     assert df.shape[0] == 15
 
-def test_convert_report_to_dataframe__cols_single_group(notifier):
-    df = notifier.convert_report_to_dataframe()
+def test_convert_report_to_dataframe__cols_single_group(email_sender):
+    df = email_sender.convert_report_to_dataframe()
     assert tuple(df.columns) == ('Termo de pesquisa', 'Seção', 'URL',
                                  'Título', 'Resumo', 'Data')
 
-def test_convert_report_to_dataframe__cols_grouped_report(notifier, report_example):
+def test_convert_report_to_dataframe__cols_grouped_report(email_sender, report_example):
     report_example['group_name_different_of_single_group'] = \
         report_example.pop('single_group')
-    notifier.search_report = report_example
-    df = notifier.convert_report_to_dataframe()
+    email_sender.search_report = report_example
+    df = email_sender.convert_report_to_dataframe()
     assert tuple(df.columns) == ('Grupo', 'Termo de pesquisa', 'Seção', 'URL',
                                  'Título', 'Resumo', 'Data')
 
-def test_get_csv_tempfile__valid_file_name_preffix(notifier):
-    with notifier.get_csv_tempfile() as csv_file:
+def test_get_csv_tempfile__valid_file_name_preffix(email_sender):
+    with email_sender.get_csv_tempfile() as csv_file:
         assert csv_file.name.split('/')[-1].startswith('extracao_dou_')
 
-def test_get_csv_tempfile__valid_file_name_suffix(notifier):
-    with notifier.get_csv_tempfile() as csv_file:
+def test_get_csv_tempfile__valid_file_name_suffix(email_sender):
+    with email_sender.get_csv_tempfile() as csv_file:
         assert csv_file.name.endswith('.csv')
 
-def test_get_csv_tempfile__valid_csv(notifier):
-    with notifier.get_csv_tempfile() as csv_file:
+def test_get_csv_tempfile__valid_csv(email_sender):
+    with email_sender.get_csv_tempfile() as csv_file:
         assert pd.read_csv(csv_file.name) is not None
 
 def test_merge_results(merge_results_samples):
